requirements.txt
# Core (required for test_pipeline.py and gp_surrogate.py + active_loop.py)
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7

# For VAE training (gb_vae.py)
torch>=2.0

# For SOAP/MBTR descriptor computation
dscribe>=2.1

# For UMAP visualisation (viz_latent.py falls back to PCA if absent)
umap-learn>=0.5

# For accessing DFT databases
ase>=3.22
